requirements.txt
openai-whisper
pysubs2
numpy